fix: small adjust to use callbacks
